+++ modules/resource-controller/src/test/java/com/ibm/cloud/platform_services/resource_controller/v2/ResourceControllerIT.java
@@ -1,5 +1,20 @@
 package com.ibm.cloud.platform_services.resource_controller.v2;
 
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertNotEquals;
+import static org.testng.Assert.assertTrue;
+import static org.testng.AssertJUnit.assertNotNull;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
 /*
  * (C) Copyright IBM Corp. 2020.
  *
@@ -12,25 +27,43 @@ package com.ibm.cloud.platform_services.resource_controller.v2;
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-
-import com.ibm.cloud.platform_services.resource_controller.v2.model.*;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.CreateResourceAliasOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.CreateResourceBindingOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.CreateResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.CreateResourceKeyOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.DeleteResourceAliasOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.DeleteResourceBindingOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.DeleteResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.DeleteResourceKeyOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.GetResourceAliasOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.GetResourceBindingOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.GetResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.GetResourceKeyOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ListReclamationsOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ListResourceAliasesOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ListResourceBindingsOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ListResourceInstancesOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ListResourceKeysOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.LockResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.Reclamation;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ReclamationsList;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceAlias;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceAliasesList;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceBinding;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceBindingsList;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceInstance;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceInstancesList;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceKey;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.ResourceKeysList;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.RunReclamationActionOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.UnlockResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.UpdateResourceAliasOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.UpdateResourceBindingOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.UpdateResourceInstanceOptions;
+import com.ibm.cloud.platform_services.resource_controller.v2.model.UpdateResourceKeyOptions;
 import com.ibm.cloud.platform_services.test.SdkIntegrationTestBase;
 import com.ibm.cloud.sdk.core.http.Response;
 import com.ibm.cloud.sdk.core.service.exception.ServiceResponseException;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.Test;
-import java.util.UUID;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.TimeUnit;
-import java.io.*;
-
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNotEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
-import static org.testng.AssertJUnit.assertNotNull;
 
 /**
  * Integration test class for the IamAccessGroups service.
@@ -66,11 +99,14 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
     String testReclamationId1 = null;
     String testReclamationId2 = null;
 
-    /**
-     * This method provides our config filename to the base class.
-     */
+    @Override
     public String getConfigFilename() {
         return "../../resource_controller.env";
+    }
+
+    @Override
+    public boolean loggingEnabled() {
+        return false;
     }
 
     /**
@@ -91,7 +127,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         assertNotNull(service.getServiceUrl());
 
         transactionId = UUID.randomUUID().toString();
-        System.out.println("Transaction-Id for Test Run: " + transactionId);
+        log("Transaction-Id for Test Run: " + transactionId);
     }
 
     @AfterClass
@@ -101,7 +137,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         if (testReclaimInstanceGuid != null) {
             cleanupReclamationInstance();
         } else {
-            System.out.println("Reclamation instance was not created. No cleanup needed.");
+            log("Reclamation instance was not created. No cleanup needed.");
         }
     }
 
@@ -1090,7 +1126,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         try {
             TimeUnit.SECONDS.sleep(20);
         } catch (Exception e) {
-            System.out.println("Failed to wait for 20 seconds.");
+            log("Failed to wait for 20 seconds.");
         }
     }
 
@@ -1122,7 +1158,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
     public void test47ListReclamationsForAccountId() {
         ListReclamationsOptions options = new ListReclamationsOptions.Builder()
             // .accountId(TEST_ACCOUNT_ID)
-            .resourceInstanceId(testReclaimInstanceGuid) //checking reclamations with instance guid for more test reliability 
+            .resourceInstanceId(testReclaimInstanceGuid) //checking reclamations with instance guid for more test reliability
             .build();
 
         Response<ReclamationsList> response = service.listReclamations(options)
@@ -1173,7 +1209,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         try {
             TimeUnit.SECONDS.sleep(20);
         } catch (Exception e) {
-            System.out.println("Failed to wait for 20 seconds.");
+            log("Failed to wait for 20 seconds.");
         }
     }
 
@@ -1216,7 +1252,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         try {
             TimeUnit.SECONDS.sleep(20);
         } catch (Exception e) {
-            System.out.println("Failed to wait for 20 seconds.");
+            log("Failed to wait for 20 seconds.");
         }
     }
 
@@ -1268,7 +1304,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
         try {
             TimeUnit.SECONDS.sleep(20);
         } catch (Exception e) {
-            System.out.println("Failed to wait for 20 seconds.");
+            log("Failed to wait for 20 seconds.");
         }
     }
 
@@ -1307,16 +1343,16 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-                System.out.println("Successful cleanup of key " + testInstanceKeyGuid);
+                log("Successful cleanup of key " + testInstanceKeyGuid);
             } catch (ServiceResponseException e) {
                 if (e.getStatusCode() == 410) {
-                    System.out.println("Key " + testInstanceKeyGuid + " was already deleted by tests.");
+                    log("Key " + testInstanceKeyGuid + " was already deleted by tests.");
                 } else {
-                    System.out.println("Failed to cleanup key " + testInstanceKeyGuid + ". Error: " + e.getMessage());
+                    log("Failed to cleanup key " + testInstanceKeyGuid + ". Error: " + e.getMessage());
                 }
             }
         } else {
-            System.out.println("Key for instance was not created. No cleanup needed.");
+            log("Key for instance was not created. No cleanup needed.");
         }
 
         if (testAliasKeyGuid != null) {
@@ -1329,16 +1365,16 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-                System.out.println("Successful cleanup of key " + testAliasKeyGuid);
+                log("Successful cleanup of key " + testAliasKeyGuid);
             } catch (ServiceResponseException e) {
                 if (e.getStatusCode() == 410) {
-                    System.out.println("Key " + testAliasKeyGuid + " was already deleted by tests.");
+                    log("Key " + testAliasKeyGuid + " was already deleted by tests.");
                 } else {
-                    System.out.println("Failed to cleanup key " + testAliasKeyGuid + ". Error: " + e.getMessage());
+                    log("Failed to cleanup key " + testAliasKeyGuid + ". Error: " + e.getMessage());
                 }
             }
         } else {
-            System.out.println("Key for alias was not created. No cleanup needed.");
+            log("Key for alias was not created. No cleanup needed.");
         }
 
         if (testBindingGuid != null) {
@@ -1351,16 +1387,16 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-                System.out.println("Successful cleanup of binding " + testBindingGuid);
+                log("Successful cleanup of binding " + testBindingGuid);
             } catch (ServiceResponseException e) {
                 if (e.getStatusCode() == 410) {
-                    System.out.println("Binding " + testBindingGuid + " was already deleted by tests.");
+                    log("Binding " + testBindingGuid + " was already deleted by tests.");
                 } else {
-                    System.out.println("Failed to cleanup binding " + testBindingGuid + ". Error: " + e.getMessage());
+                    log("Failed to cleanup binding " + testBindingGuid + ". Error: " + e.getMessage());
                 }
             }
         } else {
-            System.out.println("Binding was not created. No cleanup needed.");
+            log("Binding was not created. No cleanup needed.");
         }
 
         if (testAliasGuid != null) {
@@ -1373,22 +1409,22 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-                System.out.println("Successful cleanup of alias " + testAliasGuid);
+                log("Successful cleanup of alias " + testAliasGuid);
             } catch (ServiceResponseException e) {
                 if (e.getStatusCode() == 410) {
-                    System.out.println("Alias " + testAliasGuid + " was already deleted by tests.");
+                    log("Alias " + testAliasGuid + " was already deleted by tests.");
                 } else {
-                    System.out.println("Failed to cleanup alias " + testAliasGuid + ". Error: " + e.getMessage());
+                    log("Failed to cleanup alias " + testAliasGuid + ". Error: " + e.getMessage());
                 }
             }
         } else {
-            System.out.println("Alias was not created. No cleanup needed.");
+            log("Alias was not created. No cleanup needed.");
         }
 
         if (testInstanceGuid != null) {
             cleanupInstance();
         } else {
-            System.out.println("Instance was not created. No cleanup needed.");
+            log("Instance was not created. No cleanup needed.");
         }
     }
 
@@ -1405,7 +1441,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
 
             instance = response.getResult();
         } catch (ServiceResponseException e) {
-            System.out.println("Failed to retrieve instance " + testInstanceGuid + "for cleanup. Error: " + e.getMessage());
+            log("Failed to retrieve instance " + testInstanceGuid + "for cleanup. Error: " + e.getMessage());
         }
 
         if (instance != null && instance.getState().equals("active") && instance.isLocked()) {
@@ -1418,9 +1454,9 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                     .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                     .execute();
 
-                System.out.println("Successfully unlocked instance " + testInstanceGuid);
+                log("Successfully unlocked instance " + testInstanceGuid);
             } catch (ServiceResponseException e) {
-                System.out.println("Failed to unlock instance " + testInstanceGuid + "for cleanup. Error: " + e.getMessage());
+                log("Failed to unlock instance " + testInstanceGuid + "for cleanup. Error: " + e.getMessage());
             }
         }
 
@@ -1433,12 +1469,12 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-            System.out.println("Successfully cleaned up instance " + testInstanceGuid);
+            log("Successfully cleaned up instance " + testInstanceGuid);
         } catch (ServiceResponseException e) {
             if (e.getStatusCode() == 410) {
-                System.out.println("Instance " + testInstanceGuid + " was already deleted by tests.");
+                log("Instance " + testInstanceGuid + " was already deleted by tests.");
             } else {
-                System.out.println("Failed to cleanup instance " + testInstanceGuid + ". Error: " + e.getMessage());
+                log("Failed to cleanup instance " + testInstanceGuid + ". Error: " + e.getMessage());
             }
         }
     }
@@ -1455,7 +1491,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
 
             ResourceInstance instance = response.getResult();
             if (instance.getState().equals("removed")) {
-                System.out.println("Instance " + testReclaimInstanceGuid + " was already reclaimed by tests.");
+                log("Instance " + testReclaimInstanceGuid + " was already reclaimed by tests.");
             } else if (instance.getState().equals("pending_reclamation")) {
                 cleanupInstancePendingReclamation();
             } else {
@@ -1467,18 +1503,18 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                     .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                     .execute();
 
-                System.out.println("Successfully scheduled instance " + testReclaimInstanceGuid + " for reclamation.");
-                
+                log("Successfully scheduled instance " + testReclaimInstanceGuid + " for reclamation.");
+
                 try {
                     TimeUnit.SECONDS.sleep(20);
                 } catch (Exception e) {
-                    System.out.println("Failed to wait for 20 seconds.");
+                    log("Failed to wait for 20 seconds.");
                 }
 
                 cleanupInstancePendingReclamation();
             }
         } catch (ServiceResponseException e) {
-            System.out.println("Failed to reclaim instance " + testReclaimInstanceGuid + ". Error: " + e.getMessage());
+            log("Failed to reclaim instance " + testReclaimInstanceGuid + ". Error: " + e.getMessage());
         }
     }
 
@@ -1495,7 +1531,7 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
 
             ReclamationsList reclamationList = response.getResult();
             if (reclamationList.getResources().size() == 0) {
-                System.out.println("No reclamations for instance " + testReclaimInstanceGuid + " were returned.");
+                log("No reclamations for instance " + testReclaimInstanceGuid + " were returned.");
                 return;
             }
 
@@ -1509,9 +1545,9 @@ public class ResourceControllerIT extends SdkIntegrationTestBase {
                 .addHeader("Transaction-Id", "rc-sdk-java-cleanup-" + transactionId)
                 .execute();
 
-                System.out.println("Successfully reclaimed instance " + testReclaimInstanceGuid);
+                log("Successfully reclaimed instance " + testReclaimInstanceGuid);
         } catch (ServiceResponseException e) {
-            System.out.println("Failed to reclaim instance " + testReclaimInstanceGuid + ". Error: " + e.getMessage());
+            log("Failed to reclaim instance " + testReclaimInstanceGuid + ". Error: " + e.getMessage());
         }
     }
 }